migrate gh-cli command to gh-cli-pulls
