+++ github_api/cli/pulls/scatter_option.py
@@ -1,4 +1,4 @@
-from ..consts import Plot
+from ...pulls.consts import Plot
 
 
 def parse_scatter_argument(subparsers):
