+++ github_api/cli/pulls/violin_option.py
@@ -1,4 +1,4 @@
-from ..consts import Plot
+from ...pulls.consts import Plot
 
 
 def parse_violin_argument(subparsers):
