+++ github_api/pulls/visualization/chart.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from ..utils import log
+from ...utils import log
 
 
 def filter_dates(args, df):
